requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
fastapi==0.110.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.2.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
Werkzeug==3.0.2
